feat: authentication, user searching

+++ src/main/java/net/ddns/protocoin/controller/InvitationController.java
@@ -1,10 +1,11 @@
 package net.ddns.protocoin.controller;
 
+import net.ddns.protocoin.config.MyUserDetails;
 import net.ddns.protocoin.dto.InvitationDTO;
 import net.ddns.protocoin.dto.UserDTO;
-import net.ddns.protocoin.model.Invitation;
 import net.ddns.protocoin.service.InvitationService;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,40 +20,45 @@ public class InvitationController {
         this.invitationService = invitationService;
     }
 
-    @PostMapping
-    public ResponseEntity<InvitationDTO> addInvitation(@RequestBody Invitation invitation) {
-        return ResponseEntity.ok(invitationService.addInvitation(invitation));
+    @PostMapping("/invite/{userId}")
+    public ResponseEntity<InvitationDTO> addInvitationByIds(Authentication authentication, @PathVariable long userId) {
+        return ResponseEntity.ok(invitationService.addInvitationByIds(
+                ((MyUserDetails)authentication.getPrincipal()).getId(), userId
+        ));
     }
 
-    @PostMapping("/invite")
-    public ResponseEntity<InvitationDTO> addInvitationByIds(@RequestParam long inviterId, @RequestParam long invitedId) {
-        return ResponseEntity.ok(invitationService.addInvitationByIds(inviterId,invitedId));
-    }
-
-    @DeleteMapping
-    public ResponseEntity<Void> deleteInvitation(@RequestBody Invitation invitation) {
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteInvitationById(Authentication authentication, @PathVariable long id) {
+        var loggedUserId = ((MyUserDetails)authentication.getPrincipal()).getId();
+        var invitation = invitationService.getInvitation(id);
+        if (invitation.getUser1().getId() != loggedUserId) {
+            return ResponseEntity.badRequest().build();
+        }
         invitationService.deleteInvitation(invitation);
         return ResponseEntity.ok().build();
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteInvitationById(@PathVariable long id) {
-        invitationService.deleteInvitationById(id);
-        return ResponseEntity.ok().build();
-    }
-
     @GetMapping("/sent")
-    public ResponseEntity<List<InvitationDTO>> getSentInvitation(@RequestParam long userId) {
-        return ResponseEntity.ok(invitationService.getSentInvitations(userId));
+    public ResponseEntity<List<InvitationDTO>> getSentInvitation(Authentication authentication) {
+        return ResponseEntity.ok(invitationService.getSentInvitations(
+                        ((MyUserDetails)authentication.getPrincipal()).getId()
+        ));
     }
 
     @GetMapping("/received")
-    public ResponseEntity<List<InvitationDTO>> getReceivedInvitation(@RequestParam long userId) {
-        return ResponseEntity.ok(invitationService.getReceivedInvitations(userId));
+    public ResponseEntity<List<InvitationDTO>> getReceivedInvitation(Authentication authentication) {
+        return ResponseEntity.ok(invitationService.getReceivedInvitations(
+                ((MyUserDetails)authentication.getPrincipal()).getId()
+        ));
     }
 
-    @GetMapping("/accept/{id}")
-    public ResponseEntity<List<UserDTO>> acceptInvitationById(@PathVariable long id) {
+    @GetMapping("/{id}/accept")
+    public ResponseEntity<List<UserDTO>> acceptInvitationById(Authentication authentication, @PathVariable long id) {
+        var loggedUserId = ((MyUserDetails)authentication.getPrincipal()).getId();
+        var invitation = invitationService.getInvitation(id);
+        if (invitation.getUser2().getId() != loggedUserId) {
+            return ResponseEntity.badRequest().build();
+        }
         return ResponseEntity.ok(invitationService.acceptInvitationAndCreateContact(id));
     }
 }
